add an unique translators file for the repositories
